fix page wrapper
