+++ src/app/(pages)/layout.tsx
@@ -1,0 +1,9 @@
+import { PageWrapper } from '@/components/PageWrapper';
+
+export default function PageLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <PageWrapper>{children}</PageWrapper>;
+}
