+++ src/app/(pages)/page.tsx
@@ -1,0 +1,9 @@
+import { Stack, Typography } from '@mui/material';
+
+export default async function Home() {
+  return (
+    <Stack>
+      <Typography variant='h1'>Welcome</Typography>
+    </Stack>
+  );
+}
