+++ src/app/(pages)/places/table.tsx
