--- src/app/page.tsx
@@ -1,12 +1,0 @@
-import { PageWrapper } from '@/components/PageWrapper';
-import { Stack, Typography } from '@mui/material';
-
-export default async function Home() {
-  return (
-    <PageWrapper>
-      <Stack>
-        <Typography variant='h1'>Welcome</Typography>
-      </Stack>
-    </PageWrapper>
-  );
-}
